<!--
Thank you for creating a pull request to contribute to Hexo code! Before you open the request please answer the following questions to help it be more easily integrated. Please check the boxes "[ ]" with "[x]" when done too.
-->

## What does it do?

- update hexo-util to 1.6.1.
- include #3847 (with new `isExternalLink()`)
- update test case for `url_for`
- The test cases for `url_for()`, `full_url_for()` with `pretty_urls` configuration.

## How to test

```sh
git clone -b hexo-util-1.6 https://github.com/sukkaw/hexo.git
cd hexo
npm install
npm test
```

## Screenshots



## Pull request tasks

- [ ] Add test cases for the changes.
- [x] Passed the CI test.