@curbengh 2019-11-28
https://github.com/hexojs/hexo-util/pull/143 to be released in https://github.com/hexojs/hexo-util/pull/144