@SukkaW 2019-11-29
The test cases for `url_for()`, `full_url_for()` with `pretty_urls` configuration is now bring up.