@SukkaW 2019-11-30
The benchmark result:

| | Node.js 8 | Node.js 10 | Node.js 12 | Node.js 13 |
|:---:|:---:|:---:|:---:|:---:|
| Current master branch | 37.8s | 33.8s | 83.7s | 87.5s |
| This PR | 25.6s | 24.8s | 25.5s | 26.2s |
| PR #3834 | 27s | 27s | 27s | 26s |

The performance regression is now solved.